# mbms/backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organization/companies": {
      "get": {
        "operationId": "CompaniesController_list",
        "parameters": [
          {
            "name": "page",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "type": "number"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/organization/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "CompaniesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "CompaniesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/organization/companies/{id}/activate": {
      "post": {
        "operationId": "CompaniesController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/organization/companies/{id}/deactivate": {
      "post": {
        "operationId": "CompaniesController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/organization/companies/{id}/hierarchy": {
      "get": {
        "operationId": "CompaniesController_hierarchy",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/inventory/stock": {
      "get": {
        "operationId": "InventoryController_listStock",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lowStockOnly",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/inventory/movements": {
      "get": {
        "operationId": "InventoryController_listMovements",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "productId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/inventory/adjustments": {
      "post": {
        "operationId": "InventoryController_adjust",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdjustStockDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/inventory/products/{id}/reorder-point": {
      "put": {
        "operationId": "InventoryController_setReorderPoint",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetReorderPointDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/dashboard/companies/{id}": {
      "get": {
        "operationId": "DashboardController_companyDashboard",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
    "/api/v1/leave/balances": {
      "get": {
        "operationId": "LeaveController_listBalances",
        "parameters": [
          {
            "name": "filter[employeeId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "LeaveController_setBalance",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetLeaveBalanceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/leave/applications": {
      "get": {
        "operationId": "LeaveController_list",
        "parameters": [
          {
            "name": "filter[companyId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[employeeId]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "filter[status]",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "post": {
        "operationId": "LeaveController_submit",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitLeaveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/leave/applications/{id}/approve": {
      "post": {
        "operationId": "LeaveController_approve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/leave/applications/{id}/reject": {
      "post": {
        "operationId": "LeaveController_reject",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectLeaveDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/leave/applications/{id}/cancel": {
      "post": {
        "operationId": "LeaveController_cancel",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        }
      }
    },
    "/api/v1/leave/calendar": {
      "get": {
        "operationId": "LeaveController_calendar",
        "parameters": [
          {
            "name": "companyId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateFrom",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "Morise MBMS API",
    "description": "Morise Holdings Limited Business Management System — Phase 1 API. Base path /api/v1. Staff endpoints take a `staff` bearer token from POST /identity/auth/login; Customer Storefront endpoints take a `customer` bearer token from POST /customer-portal/auth/login.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "staff": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "customer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCompanyDto": {
        "type": "object",
        "properties": {
          "parentCompanyId": {
            "type": "string"
          },
          "relationshipType": {
            "type": "string",
            "enum": [
              "subsidiary",
              "associate"
            ]
          },
          "ownershipPercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "name": {
            "type": "string"
          },
          "registrationNumber": {
            "type": "string"
          },
          "taxId": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "financialYearStart": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "financialYearStart"
        ]
      },
      "UpdateCompanyDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "relationshipType": {
            "type": "string",
            "enum": [
              "subsidiary",
              "associate"
            ]
          },
          "ownershipPercent": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "registrationNumber": {
            "type": "string"
          },
          "taxId": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "logoUrl": {
            "type": "string"
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3
          },
          "financialYearStart": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "inactive"
            ]
          }
        }
      },
...
      "AdjustStockDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "movementType": {
            "type": "string",
            "enum": [
              "receipt",
              "issue",
              "count"
            ]
          },
          "quantity": {
            "type": "number",
            "minimum": 0
          },
          "reason": {
            "type": "string",
            "maxLength": 500
          },
          "reference": {
            "type": "string",
            "maxLength": 100
          }
        },
        "required": [
          "productId",
          "movementType",
          "quantity"
        ]
      },
      "SetReorderPointDto": {
        "type": "object",
        "properties": {
          "reorderPoint": {
            "type": "number",
            "nullable": true,
            "minimum": 0
          }
        }
      },
...
      "SetLeaveBalanceDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "leaveType": {
            "type": "string",
            "enum": [
              "annual",
              "sick",
              "maternity",
              "paternity",
              "emergency"
            ]
          },
          "year": {
            "type": "number",
            "minimum": 0
          },
          "entitledDays": {
            "type": "number",
            "minimum": 0
          }
        },
        "required": [
          "employeeId",
          "leaveType",
          "year",
          "entitledDays"
        ]
      },
      "SubmitLeaveDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string"
          },
          "leaveType": {
            "type": "string",
            "enum": [
              "annual",
              "sick",
              "maternity",
              "paternity",
              "emergency"
            ]
          },
          "startDate": {
            "type": "string"
          },
          "endDate": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        },
        "required": [
          "leaveType",
          "startDate",
          "endDate"
        ]
      },
      "RejectLeaveDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string"
          }
        }
      },
...
    }
  }
}